Update vendor archive download script for Github Actions
`make -C vendor` will download from the most-recent successful CI run for the current branch or master. There's a standalone shell script to find the CI Run, Artifact URLs, and download. GH Actions requires users to be logged in, so if `hub` isn't available, we fall back to providing URLs for the browser.