windows: use subprocess instead of exec for git
For directly passed to git commands, invoke git as a subprocess rather
than replacing the current process with git as we do on Posix.

On Windows exec() doesn't replace the current process, it spawns a
child then exits the parent, which means the child is running in the
background and the shell doesn't get any exit code. Via subprocess.run()
pipes/etc seem to stay connected correctly.