Make working_copy JSON a real object with path.
Until this change, a JSON workingCopy looked like any of the following:
null  // no working copy
{}    // working copy exists but has no uncommitted changes
{"metaChanges": ... "featureChanges": ...}  // working copy exists and
                                            // has uncommitted changes

The workingcopy has no info except changes, and so almost disappears
if there are no changes. This changes its JSON something like:

{
  "path": "buildings-project.gpkg",
  "changes": {"metaChanges": ..., "featureChanges": ...}
}

This also means the JSON is more similar to that output by commit,
which also has a "changes" field with the same format.