Expand 'show' command to display patches
`sno show` already exists, but takes no options and actually just
displays a text log entry for the given commit.
This change expands it to work much more like `git show`.

Shows a commit (default `HEAD`) as a patch, in either text
(a la `git show`) or JSON (with `--json`).

HTML and 'quiet' output could be added in future.

The patch is a diff plus some meta-information. The diff
is just the same as the `sno diff` output for the relevant
commit.

In text mode, the meta-information is the same as what git show outputs:

```
commit ad9797ea83000864c2ab98c421d2d5256a12db8f
Author: Craig de Stigter <craig@destigter.nz>
Date:   Wed Apr 15 13:19:16 2020 +1200

    commit

<diff output follows>
```

In JSON mode, the metadata is added to the top level JSON object
so that the patch is also a diff. It looks like:

```json
{
  "sno.diff/v1": {...},
  "sno.patch/v1": {
    "authorName": "Craig de Stigter",
    "authorEmail": "craig@destigter.nz",
    "authorTime": "2020-04-15T01:19:16Z",
    "authorTimeOffset": "+12:00",
    "message": "commit"
  }
}
```