Refactor diff.py so RepositoryStructure neednt contain a commit.
I wrote a change so that RepositoryStructures could point at any
particular tree, not just at any particular commit.
This could be useful for doing diffs against the empty tree, which
could be used to `sno show` the first commit in a repository.
(sno show always delegates to sno diff, and the first commit cannot
be diffed right now since there are no other commits to diff against.)

Or it could possibly be useful to do arbitrary diffs between
any two repository structures, without reference to their
shared history.

However, leaving the usefulness of that change aside for a moment:

The refactoring required to do this turned out to be a nice
little refactoring on its own, so I've broken it off into its own
change - which you see here. Allowing diffs of trees as well as
commits can be discussed separately.