Better mutually exclusive options
Fixes #99.

The existing MutexOption class only works where the options are given
different names. For most of the cases we're using it, that isn't the
case - the options all have the name `output_format`. By the time the
`MutexOption.handle_parse_result` is called, the options have been
collapsed into one (`{"output_format": "json"}`) which is the last
option to have been specified on the command line.

This change adds a way of handling the options together via a grouping
decorator. The options are parsed individually, but then checked for
mutual exclusion and only one value passed to the command function.