MutexOption: fix a bug and improve error msgs
Before this, imports with `--list` or `--all-tables` failed with:

```
$ sno import ./census2016_sdhca_ot_short.gpkg --list abc
Error: Illegal usage: 'do_list' is mutually exclusive with tables.
```

This happened even if no tables were specified, because `tables` is an
argument so it was still in the context:

```
$ sno import ./census2016_sdhca_ot_short.gpkg --list
Error: Illegal usage: 'do_list' is mutually exclusive with tables.
```

That's now fixed, and also the error output makes more sense (uses the
option names instead of the internal variable names):

```
$ sno import ../gpkg-au-census/census2016_sdhca_ot_short.gpkg --list abc
Error: Illegal usage: '--list' is mutually exclusive with '[TABLES]...'.
```