Speed up feature iteration using memoryviews
This is a ~10% speedup for feature-iteration commands (eg checkout)

`blob.data` is slow, partly because it copies the data into a
bytestring. In most cases we then pass the bytestring to msgpack, and
then throw it away afterwards.

However, blobs support the buffer protocol, so we can cast them to
memoryview instead. msgpack accepts memoryviews. This minor change
avoids the copy and thus gives us a sizeable speedup.