Normalise geometries from WC in datasetV2
Clients won't produce GPKG geometries to our specification,
so we need to clean them when committing or diffing.

Can't do this sanely for V1 datasets because we don't know the exact
form of the geometries that already exist in the repos.
So we leave this as a datasetV2 only feature