commit/apply: Faster large commits
Background: We were using Index objects to create trees from diffs.

Simply, we'd:
 - create an Index
 - read the full tree of the repo into it
 - call .remove() and .add() a bunch of times to apply a diff
 - write the index back into a tree
 - create a commit object with the new tree.

Unfortunately it turns out this performs poorly at scale, because
.remove() and .add() spend an awful lot of time doing memmove()
operations.
It turns out that indexes seem to be basically sorted vectors, so
inserting and removing is slow.

This change replaces indexes with TreeBuilder objects.
They have a somewhat lower-level API. each TreeBuilder represents
a single directory in a tree. So to modify a large swath of subtrees,
you have to use treebuilders to create each tree, and then add each
tree to it's parent tree.

This requires quite a bit of recursive support code so we can write
features and meta blobs. The result is an 80-90% reduction in time
to apply large patches (10000+ small features. I haven't tested
large features) and the same will be true of regular large commits.