Make test_apply_benchmark test actually useful
Background: I noticed the test_apply_benchmark test is basically
useless because it creates a patch in a silly way.
It does an UPDATE to all rows in the WC, then commits it, then creates
a patch with create-patch, then applies the patch to another branch.

But of course, by the time it applies the patch, all the blobs/trees
already exist, so the apply is super fast.

This change just puts the patch in a fixture and makes the benchmark
more useful.