apply: Add --allow-missing-old-values option
Makes it easier to generate patches in an external system.

Specifically, #283 has meant some of our internally mirrored repos
have gotten into a state where some meta blobs may or may not exist,
and it's difficult to externally generate a patch without knowing
whether they exist or not.

This change allows our external system to just not supply the "-"
object for each change, and sno will treat the delta as either an insert
or an update as necessary.

For external systems which can be confident in the state of the repo,
this is useful for performance also:

 * patches can be half as large, since old revisions aren't needed
 * patch application can be much faster, since we don't need to fetch
   and decode the old features for every delta.

Even if this flag is specified, any "-" values supplied in the patch
will be checked for conflicts. It only skips checking missing ones.