init: Change default branch to 'main'
'main' is likely to be the default in the git project shortly,
and makes sense to incorporate the same change in sno as soon as
possible to avoid further disruption as usage increases.

Because sno relies on git's config mechanism and we want to use
the existing `init.defaultBranch` config variable,
this change to the default value of that variable relies on overriding
it iff it hasn't been set.

Only 'init', 'clone' and 'config' commands actually use that variable,
so this change only parses the config and overrides that value at
startup in those three commands, to avoid slowing down startup for
other commands.