import: Add --replace-ids flag
The existing `--replace-existing` flag imports over the top of an
existing dataset.

Imports from huge source tables can be slow, since the entire table
must be read and imported into git, even if only a small subset of the
table changed.

If the caller knows what parts of the table have changed, it makes
sense for them to supply that information to sno, to speed up the
import.

This change adds a `--replace-ids` flag, which supplies a list of
primary key values which sno should look at when determining the
changeset to import.