pyodbc/mssql: platform packaging improvements
macOS/Linux:

The MSSQL ODBC drivers depend on unixodbc, and will install the system
unixodbc package via Homebrew/Deb/RPM/etc. This means we end up with
two libodbc libraries loaded in the same process, which is bad. And
we've run into libc/libstdc++ issues since we're building on an old
libc but the MSSQL drivers are built separately for each OS release.

So on Linux we build pyodbc against the system library, and on macOS
we use the binary wheel from PyPI (which links against Homebrew too).
Then we explicitly don't delocate/fix the python module, so it'll always
refer to the system library. Then we need to handle the ImportError in
Python if pyodbc can't be loaded because unixodbc isn't installed.

PyInstaller tries to be helpful, so disable that behaviour too.

On Windows: switch to using the binary wheel from PyPI.