Fix locking regression: only change journal mode from DELETE -> TRUNCATE
The default GPKG journal mode is DELETE, the only mode that persists is WAL.
Trying to change the journal mode from WAL -> DELETE has 2 problems -
specifically, there are 2 reasons the GPKG journal mode might be WAL,
and not the default.

1. The user might have changed it manually to WAL. In which case,
we should respect their decision.
2. QGIS automatically changes the journal mode to WAL. This may
or may not have been a good design decision by QGIS, but, they
only switch the journal mode to WAL for as long as QGIS is connected,
and, it is not possible to change journal mode while more than one
connection is open. In which case, if we try to change the mode
to TRUNCATE, it will crash. So, we shouldn't touch it in this case
either.

This logic existed in Sno 0.7.1 - I lost it during the port
to sqlalchemy, and unit tests / end to end tests don't pick up on it
because we don't simulate QGIS accurately (ie, we don't maintain a
QGIS-like connection and put the journal mode in WAL while we run
Sno commands).