`status`: Faster with many working copy changes
This commit avoids generating a DeltaDiff for feature changes, and
instead just counts the features for updates/deletes/inserts change
types. The diff was only used to do rename detection.

caveat: As a result `status` no longer detects renames (updates to the
PK field) which makes it slightly inconsistent with `diff`.