postgis WC: Avoid INSERT to spatial_ref_sys
Unless user is doing something rather exotic with custom CRSes, there's
no need to write to spatial_ref_sys. The existing code acknowledges
this, but still does an INSERT (with a fallback to ON CONFLICT UPDATE)

However doing an INSERT requires additional privileges beyond what might
be expected. We only need read-only access to spatial_ref_sys
for normal operation.

This commit does a SELECT first to determine if an INSERT is really
necessary, so that in most cases we don't need INSERT perm on the table.