diff: Add a way to get diff size estimates
Large diffs are currently quite slow. Even if we can optimise a lot,
it's not uncommon to diff an import commit and produce 50M features
of output, which can take minutes/hours to generate and be
difficult or impossible to handle.

This change adds a size estimate feature so scripts can avoid showing
large diffs altogether.

The feature is implemented via a flag:

```
  --only-feature-count [veryfast|fast|medium|good|exact]
                                  Returns only a feature count (the number of
                                  features modified in this diff). If the
                                  value is 'exact', the feature count is exact
                                  (this may be slow.) Otherwise, the feature
                                  count will be approximated with varying
                                  levels of accuracy.
```

`fast` is normally good enough for a ballpark estimate.