Disable deltified objects by default
This is a continuation of #159 and #163,
in which we discovered that even a small proportion of deltified
objects (~0.1%) with small delta chains can cause catastrophic
performance issues for diff. We've since discovered it causes similar
problems with clone, if the deltas exist on the server.

sno's repo layout is quite different than your typical git repo
(there are typically millions of small blobs, distributed much more
evenly across a constant number of trees) and no deltas seems to be
a better fit for large sno repo layouts.

Set `pack.window` and `pack.depth` in your gitconfig explicitly to
override this default.