Enable auto-incrementing PKs for all WC types.
This is normally simply a convenience for the user - the user
doesn't have to think of a new PK, the next unassigned one is
automatically supplied for them.

However, it serves another purpose when a spatial filter is active.
The PK that is automatically supplied is guaranteed not to clash
with any feature inside OR outside of the spatial filter - the user
may have no other easy way of providing such a PK themselves without
disabling the spatial filter.