Make `kart create-workingcopy --delete-existing --discard-changes` a bit more powerful
Currently it tries to check if there are uncommitted changes in the WC,
but regardless of the result, it deletes and recreates the WC.
The check for uncommitted changes can fail if the WC has been corrupted
in some way, which causes this command to fail - in exactly the situation
where this command would be ideal to rewrite the WC from scratch to fix
the corruption. This change means we don't check for uncommitted changes
if we plan to discard them either way, so the command won't fail for
that reason.