ci: macos: code-signing build fixes
Homebrew seems to _sometimes_ have signed binaries. This happens in GH
Actions CI but not locally.

And despite `codesign`s documentation saying that "existing signatures
will not be replaced, and the signing operation fails." it doesn't
actually fail, at least when using --deep on an App bundle.

Potentially we mess with dylibs to rewrite rpaths/etc so existing
signatures may be invalidated anyway. Strip any signatures from existing
files in the macOS vendor bundle when we assemble it.