Loosen dataset path restrictions
The rules for dataset paths added in #501 were a bit too unwieldy and
hard to follow.
Since we didn't have any clear reason for forbidding initial characters
in dataset paths, we discussed and decided to relax the restriction, so
that now the first character of a dataset path has almost the same
restrictions as the second and later characters. (However, starting with
a `/` or `.` is still forbidden)

Additionally this change fixes an unintentional behaviour in `import`
where a `.` in a dataset path was replaced with a `/`. This was intended
to import schema-qualified database names in a hierarchy to avoid name
conflicts when importing with `--all-tables`.

However we felt that was a misfeature; the user probably doesn't want to
import schema-qualified names, and if they do they can do so explicitly.
So this change removes that behaviour also.