log: Fix limit options not working
In https://github.com/koordinates/kart/pull/498 I added some code to
pre-parse arguments that we're going to pass to `git log`, to figure out
which of the arguments were paths.

This was necessary for future work
(https://github.com/koordinates/kart/issues/496). Unfortunately I failed
to account for git's options that take an argument, notably `-n 1`, and
the `1` got treated as a path.

This change fixes the bug for:
* `-n / --max-limit`
* `--after` / `--since` / `--until` / `--before`
* `--skip`
* `--author` / `--committer`
* `--grep`

And incidentally adds those options to the `--help` output.