Don't modify the config when fetching missing blobs
Fetching missing blobs using the remote's URL adds a new entry to the
config, whereas fetching missing blobs using the remote's *name* reuses
the existing entry in the config (ie, it looks up the remote by its
name). We don't want to modify the config, so we should use the remote's
name always.