Disallow expanding the partial clone filter (for now)
Changing the spatial filter can be done at any time. This changes the
user's view of the repo. However, currently the partial clone spatial
filter can only be set once: during clone. This is a bit of limitation in
how partial clones work, but we may be able to fix it later.
Since we can't currently change the partial clone size, we shouldn't allow
the user to reconfigure the spatial filter to be bigger than the data
they cloned (until we fix it).
Note that partial clones are still turned off unless you enable them
using the experimental environment variable, so this doesn't affect
anyone yet - but we can't turn on partial clones without either
enforcing this restriction, or fixing it.