Covert both points and poly-lines/-gons via envelopes
Previously poly-lines/-gons were converted to envelopes, then transformed,
then converted to S2 cells, whereas points were transformed and converted
to S2 cells directly. However, this special casing of points just made
the code more compliated without any actual benefit, so this is now
removed and the two stage nature - geometry -> envelope,
envelope -> S2 cells - is made more explicit.

Apart from tidying the code, this could be the first step to switching
the s2-index to just be an index of envelopes if we decide to do that.