Indexing: skip features if transformed envelope is invalid
This can happen if the CRS is changed at some point - right now we try
to apply all CRS that have ever been associated with a dataset to each
feature, which can result in invalid transforms. If this fails, then
skipping the feature means it matches every spatial filter, which is
suboptimal in terms of bandwidth, but will still work fine.