`diff -o json-lines`: Add feature count estimate
With new `--add-feature-count-estimate=<accuracy>` option, adds a line
to the output stating the number of features expected to be in
the diff. This should be useful for showing a progress bar when the diff is
large.

For small diffs it may never appear, as the estimate can easily take
longer to calculate than the whole diff anyway.