Check if libgit2 library is the right one in CMAKE
The CMAKE build looks for an already installed libgit2 library. However,
we require that libgit2 has a few changes that haven't yet been merged to
libgit2/libgit2 - this checks to make sure that those changes are present,
and instructs the user how to install our libgit2 fork if they aren't.