Ensure conflicts loaded during `kart conflicts`
Post merge, some of the conflicts may not be present on the repo due
to a spatially filtered clone. So, we need to fetch them before displaying
them.

(This won't happen usually since usually, the user won't be merging edits
that are outside their spatial filter. But it can happen.)