Start of point cloud import
Imports individual point cloud tiles by writing pointer files and
putting the tiles in the LFS cache.
This is sufficient so that running `git push` or `kart push` will
cause the tiles to be written to the remote's LFS server (if LFS is
installed).

Lots more things to be done in subsequent changes:
- No point-cloud dataset metadata is written
- Point cloud working copy is not checked out
- Although we generate the necessary tile metadata, we don't yet write
  it to the pointer files since Git LFS doesn't currently allow this