Fix slowness when checking out spatial-filter WC
Slowness is due to libgit2 repeatedly refreshing whenever it encounters
a missing object, in the hopes that the missing object is only due to
its stale view of the data on disk. The fix is to refresh once before
the WC checkout and then disable refreshing for the rest of the
checkout - this current-as-of-start-of-checkout view of the ODB is
good enough for us since we don't read and write concurrently.