Renames version to table-dataset-version
The repo version is renamed to make clear it is no longer the
repo version - instead it is the repo-wide, table-dataset version.
Table-datasets are to be unique in that they will have a repo-wide
version marker. Other datasets will be responsible for storing their
own version identifiers.

Part of the foundation for point clouds (and other dataset types).