Support filtering datasets by type
Now that there begins to be more than one type of dataset -
table and point-cloud - we need to be able to iterate
over just the table datasets or just point-cloud datasets.

This makes database working-copy checkout only happen for
table-datasets, but stops short of filtering all commands
to only look for table-datasets - as a result, commands
like `kart diff` won't work in repos with point-cloud
datasets until this functionality is implemented.