Add commit<>commit diffs to point-cloud datasets
Moves the feature_diff code out of the table-based datasets and into
the base-dataset. This code runs a git-diff, then takes all the files
that have changed and optionally uses a method to decode the paths
(eg decode_to_primary_key) and a method to transform the blob contents
(eg get_feature_from_blob). Similar but more basic transforms now
exist for point-cloud tiles that make a summary of the tile just from
the information in the pointer file.