add ext-run command for extension invocation
It can be useful to run scripts inside Kart's execution environment for
prototyping or extensions.

Add a new command, `kart ext-run PY_SCRIPT [ARGS...]` to do this.

The script's path is appended to sys.path so it can do imports of
sibling scripts.

Kart currently doesn't have a public API, so is very much user beware.