Store PC metadata in LFS pointer files
Allows for better tile diffs including commit<>commit diffs,
since we can turn pointer-file change into a human-readable tile-change,
regardless of whether we actually have the PC tile cached at present.
See improved diff output at test_point_clouds.py

Currently squashes metadata into an LFS extension - this can change if
we find a better solution / modify git LFS.

The actual metadata that we store is subject to change, but this
gets the pieces in place so that we can do so.

Another tweak:
Always store the CRS in "crs.wkt" - this is simpler than the current
system, which stores it by name, which was needed for vectors but adds
unnecessary complexity here.