Demote tabular working copy to be a "part" of the WC
Right now in Kart "working copy" is synonymous with "tabular working copy".
We need working copy to encompass more than this in order to support
point clouds. This is the first in a series of changes that make the
tabular working copy just one "part" of the working copy, and add
support for doing operations or checking the validity of all parts
simultaneously.

Currently leaves lots of TODOs in the code, where the code still
interacts directly with the tabular working copy but will have to
eventually operate on all parts simultaneously.