Allow missing WC parts.
Work on making Kart more lenient if some part of the WC are missing when
it tries to do a WC diff, since it will soon be common for only
part of the WC to be present (ie a repo with only tabular data
generally has no PC WC, and a repo with only PC data may have
no tabular WC).

Currently has no real effect since there is only one part that is
implemented - the tabular part - but is part of the tidy up /
refactor for implementing the PC part too).