More working copy parts refactoring
Access to working copies should generally go through
the repo.working_copy interface - that way more parts can
be added transparently, since operations can be performed
on all parts at once. This fixes most calls to go via this interface.

Various clean ups:
Got rid of reset_wc_if_needed - its the caller's responsiblity
to decide when and how to reset the WC.
(This was of limited utility anyway since the caller in any case has to
decide whether to check for uncommitted changes and warn the user, etc).

No need for "rewrite_full" - this is just a variation on reset().
No need for "_add_datasets_to_working_copy" - this is also just a reset() operation.