Fix kart branch -o json for empty repos
Fix `kart branch -o json` when branch HEAD is unborn
 or repo is empty - this would either fail outright
or report that HEAD is on branch "null" which is not exactly true.
When HEAD is unborn, then the HEAD is still on a branch, sort of,
but that branch is unborn. We can still report the name of the
branch, which is more useful than not doing so.