Formalise post-commit WC operations
After committing we need to do any/all of the following:
- update the WC to be tracking the latest commit-tree
- clear some PKs from the tracking table so they are no longer marked
  as dirty
- drop some features from the WC if they are outside the spatial filter

This adds that compound-operation to the WC, and tidies up the names
etc of some of the sub-operations (eg mark_as_clean is clearer than
reset_tracking_table). It also switches these methods to take key_filters
instead of ad-hoc structures like dicts or lists, which makes them
a little more self-documenting.