Add tab completion for kart
Supports completion for `refs`, `OutputFormatType`, conflicts and
dataset paths. To use this commit one needs to setup shell
completion using click [1]:

        eval "$(_KART_COMPLETE=<shell_name>_source kart)"

For example: to setup click completion for bash:

        eval "$(_KART_COMPLETE=bash_source kart)"

The existing context in Click's shell complete method creates a different
context than what's passed to the commands.Due to this the `repo` in
`ctx.obj` does not exist in the shell complete context. Hence,
a new context is created for tab completion in Kart.

[1] https://click.palletsprojects.com/en/8.1.x/shell-completion/