Fix shapefile import
- Don't use FID as a primary key, don't do anything with FID
- Use any other field as the primary key if required, otherwise,
  generate a new primary key.
- Make tests check both of these behaviours
- Delete tests that convert GPKG to SHP files to use as test data -
  checked in equivalent test data. The auto-converted test data
  was bad anyway: it had lost the primary key columns.

Fix the PkGeneratingTableImportSource - it shouldn't modify the
schema of the delegate import-source that it wraps. (Modifying
the schema of an import-source in place is allowed, as long
as you only change the IDs of the columns - these IDs are
Kart-internal and so are arbitrary for an import-source until
and unless we align them to an existing table. But, other
modifications to the import-source schema - such as adding
a new `auto_pk` column - can cause the import-source to crash,
so this must be done at the outer wrapper class only - in
the PkGeneratingTableImportSource).