Start of PC meta changes - store format.json
If we want the file type to be a constrained part of the dataset,
we should store what it is and show diffs when it changes.
This standardises the different flavours of file types as a "meta item",
which is called "format.json" and looks like so:
{
  "compression": "laz",
  "lasVersion": "1.4",
  "optimization": "copc",
  "optimizationVersion": "1.0",
  "pointDataRecordFormat": 7,
  "pointDataRecordLength": 36
}

Note that not all of these will necessarily be constrained for all
datasets - in particular, it is allowed to have a dataset with both
COPC and non-COPC files. If some of these fields are not constrained,
they will be omitted from format.json.

Individual tile diffs will show the format of individual tiles, since that
can differ slightly even as the dataset format remains unchanged - eg
the user might commit new COPC or non-COPC tiles, and being able to see
which is which in the diff could be useful.

Also: since all of this metadata - whether we store in in JSON files,
or encode it into pointer files - forms part of our API since we
expose it in diffs, I've updated the names to be JSON-like everywhere.
This is consistent with our existing diffs - schema.json has a field
"dataType" not "data-type" or "data.type".

- We could use different names internally eg "extent.crs84" internally
and then switch it to "crs84Extent" externally, but there's no point
(and like Git we tend to leak our internals by exposing lower-level
commands for hacking them if needed).
- We could use differnt names for text based diffs eg "crs84Extent"
for JSON diffs and "CRS-84 Extent" for text diffs, but again there's
no point teaching user's two different names for the same thing.