PC Meta diffs
Detect changes to PC meta-items due to changes in PC meta-tiles,
add them to diff output. They can't currently be committed.

Codify something called a ListOfConflicts. Sort-of exists already
for metadata.xml in GPKGs, but it is now  much easier to create
a ListOfConflicts with Point Clouds, so it becomes a named Kart
concept that is easier to detect and outputs properly in diffs.

Fixed schema.py to support PC schemas that don't have an "id".

Still TODO:
- make PC diffs committable (when they don't have conflicts)
- improve the error messaging around ListOfConflicts
- more logic around formats (eg, allow committing a COPC dataset to
  a --preserve-format dataset)
- maybe improve the output of a ListOfConflicts in JSON