Add basic support for --convert-to-dataset-format
This flag is now supported by PC datasets when generating WC diffs.

`kart commit` has partial support of this flag, but only insofar
as it respects it while it generates the high-level diff of what is
to be committed - it doesn't yet actually perform the in-place
conversion of the PC tiles in the WC that need to be converted,
instead exiting with NOT_YET_IMPLEMENTED if it encounters such a tile.