Store sourceOid if it differs from oid
Kart diffs which involve format conversions already have two different
oids - "sourceOid", the hash of the input file, and "oid", the hash
of the file as committed.

This change stores the sourceOid in the LFS pointer file so that
subsequent committers can avoid recommitting the same file again -
if an input file matches *either* sourceOid or oid, there's no point
recommitting it. The plan is to add this logic to point-cloud-import.