point-cloud-import: --amend
The last missing piece of tile-by-tile imports is being able to easily
amend the previous commit with a new tile.

This was tricky; I couldn't find a way to do so in git-fast-import
itself as it isn't really designed for this purpose. You can't create a
commit with a custom starting tree.

So instead, we:
* add a custom ref, do the fast import onto that ref
* create a detached commit using pygit2 which squashes the fast-imported
  commit into the original head commit
* finally, move the head branch tip to point to the new amended commit.