Fix workdir diff behaviour wrt mtimes
Implements and adds a test the following:
1. Diffs work properly when files have mtimes (modified-timestamps)
that make it look like the file has been modified, but in fact it has not.
2. Running a diff causes the mtimes to be updated for unmodified files,
where the mtimes in the index no longer match the actual file. (This
means the next diff can run quicker since we can use the mtime check instead
of the comparing hashes, which involves hashing the file and takes longer).