Support conflict resolution using the working copy
Adds --with=workingcopy to `kart resolve` in which case,
Kart will take the feature or tile from the WC as the resolution
to the specified conflict.

Note that the WC is not currently not kept in sync with the
merge - it continues to track the branch HEAD labelled "ours".
Nor is there currently a way to actually see the merge-conflicts
of PC tiles. This means the following is still needed in further PRs:

- keep the WC up to date with the merge as conflicts are resolved
- require that meta conflicts are resolved before other types of conflicts
- write all versions of conflicting tiles to the working copy, remove
them as they are resolved
- clean up any merge-state that remains - files or folders containing
conflicts - when the merge state is exited using `--continue` or `--abort`