Avoid isinstance checks
There's a few places where we check isinstance(x, pygit2.Commit)
and similar. These can be replaced by checking x.type or x.type_str -
doing so makes it easier to sub in alternative implementations of
commits, trees, or blobs.
This is probably not necessary after all right now but is harmless
and a slightly better pattern than the isinstance checks.