Remove bare-style vs tidy-style distinction
In 2020, "non-bare" Kart repos were built on top of bare Git repos.
We added support for a new type of Kart repo that was built on top
of a non-bare Git repo, thus hiding away all the internals.
This was known as "tidy style" and is now used exclusively for
non-bare Kart repos, replacing the old "bare style" non-bare repos.

So, now there are again only two types of Kart repos -
bare (same as always) and non-bare ("tidy style").