Add a working-copy merger that updates the WC during a merge
In particular, this writes all versions of a tile conflict to the
working copy, but it also writes objects (meta items, features, tiles)
to the working copy if they merge cleanly - so that the user can
inspect those things, but also for consistency: it doesn't make sense
to write only the conflicts to the WC and leave everything else as
"ours" version for simply since that how it was before the merge.

Still TODO in further PRs:
- Further update objects in the WC as they are resolved, including
tidying up any (ancestor / ours / theirs) conflicting versions.
- Use the as-merged-so-far version of the dataset to serialise any
resolves - this is relevant for feature conflicts in the case that
the schema changes during the merge.