Rename MergeIndex to MergedIndex
We have more and more context surrounding a merge.
Nice to distinguish those items that are related somehow
other HEAD that we merge with - eg MERGE_HEAD - vs those things
that are related somehow to the merged-state, after the merge -
eg MERGED_INDEX (and soon, MERGED_TREE which is the same but a tree).