Tidy up exec.py, rename to subprocess.py
We never exec anyway so _KART_NO_EXEC is confusing.
But we do still need to do subprocess.run slightly differently during
tests, unless we figure out a better way to capture output.

This makes the names of things more in line with how they work,
and makes calls to kart/subprocess.py look and work identically to
calls to subprocess (until we turn on _KART_TEST).

(Ideally we would get rid of _KART_TEST and use fd-capturing,
but this will take a bit more expertise. However, this still gets
us a tiny bit closer to that goal).